refactor: 불필요한 코드 제거 및 폴더구조 변경
