+++ src/components/CardEditor/CardDueDateInput/index.jsx
@@ -1,7 +1,7 @@
 import PropTypes from "prop-types";
-import InputBox from "components/common/InputBox";
-import { Input } from "components/common/Input/style";
 import { CardInputWrapper } from "pages/style";
+import InputBox from "components/InputBox";
+import { Input } from "components/Input/style";
 import { NAME } from "constant";
 
 function CardDueDateInput({
