+++ src/components/CardEditor/CardNumberInput/index.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
-import InputBox from "components/common/InputBox";
-import { Input } from "components/common/Input/style";
+import InputBox from "components/InputBox";
+import { Input } from "components/Input/style";
 import { CardInputWrapper } from "pages/style";
 import { CARD_NUMBER } from "constant";
