+++ src/components/ColorBox/style.js
