+++ src/components/Modal/style.js
