+++ src/components/Palette/index.jsx
