+++ src/components/Palette/style.js
